requirements.txt
mysql_connector_repackaged==0.3.1
tika==2.6.0
transformers==4.39.3
